Moved createMapChart into d3 folder
