+++ src/components/MapChart.js
@@ -1,4 +1,4 @@
-import createMapChart from './viz/createMapChart'
+import createMapChart from './d3/createMapChart'
 import withResponsiveness from '../higherOrderComponents/withResponsiveness'
 import createChartComponent from '../higherOrderComponents/createChartComponent'
 import compose from '../util/compose'
